+++ ETLBox/src/Definitions/Tasks/DataFlowTask.cs
@@ -22,35 +22,6 @@ namespace ALE.ETLBox {
             }
         }
 
-        private IDbConnectionManager _sourceDbConnection;
-        private IDbConnectionManager _destinationDbConnection;
-
-        public IDbConnectionManager SourceDbConnection
-        {
-            get {
-                if (_sourceDbConnection != null)
-                    return _sourceDbConnection;
-                else if (this.ConnectionManager != null)
-                    return (IDbConnectionManager)ConnectionManager;
-                else
-                    return DataFlow.DataFlow.SourceDbConnection;
-            }
-            set => _sourceDbConnection = value;
-        }
-        public IDbConnectionManager DestinationDbConnection
-        {
-            get
-            {
-                if (_destinationDbConnection != null)
-                    return _destinationDbConnection;
-                else if (this.ConnectionManager != null)
-                    return (IDbConnectionManager)ConnectionManager;
-                else
-                    return DataFlow.DataFlow.DestinationDbConnection;
-            }
-            set => _destinationDbConnection = value;
-        }
-
         public virtual int ProgressCount { get; set; }
 
         public bool HasLoggingThresholdRows => LoggingThresholdRows != null && LoggingThresholdRows > 0;