+++ ETLBox/src/Toolbox/DataFlow/RowTransformation.cs
@@ -78,6 +78,7 @@ namespace ALE.ETLBox.DataFlow {
             this.TaskHash = task.TaskHash;
             this.TaskName = task.TaskName;
             this.TaskType = task.TaskType;
+            this.DisableLogging = task.DisableLogging;
         }