+++ ETLBoxTest/src/Helper/TestHelper.cs
@@ -1,6 +1,7 @@
 ﻿using ALE.ETLBox;
 using ALE.ETLBox.ConnectionManager;
 using ALE.ETLBox.ControlFlow;
+using ALE.ETLBox.DataFlow;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Linq;